chore: restructure src code
