--- eventsourcingdb/errors/invalid_parameter_error.py
@@ -1,7 +1,0 @@
-from .custom_error import CustomError
-
-
-class InvalidParameterError(CustomError):
-    def __init__(self, parameter_name: str, reason: str):
-        super().__init__(
-            f'Parameter \'{parameter_name}\' is invalid: {reason}')
