--- eventsourcingdb/handlers/write_events/__init__.py
@@ -1,3 +1,0 @@
-from .preconditions import Precondition, \
-    IsSubjectPristinePrecondition, \
-    IsSubjectOnEventIdPrecondition
