+++ eventsourcingdb/is_heartbeat.py
