+++ eventsourcingdb/is_stream_error.py
