+++ eventsourcingdb/observe_events/observe_from_latest_event.py
