+++ eventsourcingdb/read_event_types/__init__.py
@@ -1,0 +1,2 @@
+from .event_type import EventType
+from .is_event_type import is_event_type
