+++ eventsourcingdb/read_events/__init__.py
@@ -1,3 +1,4 @@
 from .read_events_options import ReadEventsOptions
 from .read_from_latest_event import ReadFromLatestEvent
 from .if_event_is_missing_during_read import IfEventIsMissingDuringRead
+from .order import Order
