+++ eventsourcingdb/read_subjects/__init__.py
@@ -1,0 +1,1 @@
+from .is_subject import is_subject
