+++ eventsourcingdb/read_subjects/is_subject.py
