+++ tests/conftest.py
@@ -1,5 +1,5 @@
 import pytest_asyncio
-from eventsourcingdb.event.event_candidate import EventCandidate
+from eventsourcingdb import EventCandidate
 from .shared.database import Database
