+++ tests/test_observe_events.py
@@ -2,13 +2,12 @@ import asyncio
 from aiohttp import ClientConnectorDNSError
 import pytest
 
-from eventsourcingdb.errors.server_error import ServerError
-from eventsourcingdb.event.event_candidate import EventCandidate
-from eventsourcingdb.handlers.bound import Bound, BoundType
-# pylint: disable=C0301
-from eventsourcingdb.handlers.observe_events.if_event_is_missing_during_observe import IfEventIsMissingDuringObserve
-from eventsourcingdb.handlers.observe_events.observe_events_options import ObserveEventsOptions
-from eventsourcingdb.handlers.observe_events.observe_from_latest_event import ObserveFromLatestEvent
+from eventsourcingdb import ServerError
+from eventsourcingdb import EventCandidate
+from eventsourcingdb import Bound, BoundType
+from eventsourcingdb import IfEventIsMissingDuringObserve
+from eventsourcingdb import ObserveEventsOptions
+from eventsourcingdb import ObserveFromLatestEvent
 
 from .conftest import TestData
