+++ tests/test_read_subjects.py
@@ -1,8 +1,8 @@
 from aiohttp import ClientConnectorDNSError
 import pytest
 
-from eventsourcingdb.errors.server_error import ServerError
-from eventsourcingdb.event.event_candidate import EventCandidate
+from eventsourcingdb import ServerError
+from eventsourcingdb import EventCandidate
 
 from .conftest import TestData
